StockSim/requirements.txt
django-cryptography
django-environ
django-cron
django-simple-history
django-adminlte3
django-encrypted-model-fields
django-utils-six

Cython
numpy
pandas
matplotlib
pymysql
pyrh

Scikit-learn
future
mock
h5py
keras_preprocessing
keras_applications
gast
futures
protobuf
pybind11

Pillow
pyEX

# sudo pip3 install --pre --extra-index-url https://developer.download.nvidia.com/compute/redist/jp/v44 tensorflow
# pip3 install git+https://github.com/LichAmnesia/Robinhood.git